Move product code to a seperate file from amazon
Since there will be different api sources, move product saving to a
seperate area.
